add ability to display dialogue in game container
