+++ assets/js/dialogue.js
@@ -1,28 +1,35 @@
 let dialogueQueue = [];
 let dialogueIndex = 0;
+let appendToGameContainer = false;
 
-const showDialogue = (dialogue) => {
+const showDialogue = (dialogue, appendToContainer = false) => {
   dialogueQueue = dialogue;
   dialogueIndex = 0;
+  appendToGameContainer = appendToContainer;
   updateDialogue();
 };
 
 const updateDialogue = () => {
+  const parentElement = appendToGameContainer
+    ? document.getElementById("game-container")
+    : document.body;
   const dialogueBox = document.getElementById("dialogue-box");
-  const dialogueText = document.getElementById("dialogue-text");
-  const choicesContainer = document.getElementById("choices-container");
 
   if (dialogueIndex < dialogueQueue.length) {
     dialogueBox.classList.add("active");
+    const dialogueText = dialogueBox.querySelector("#dialogue-text");
+    const choicesContainer = dialogueBox.querySelector("#choices-container");
+
     dialogueText.innerText = dialogueQueue[dialogueIndex].text;
 
-    choicesContainer.innerHTML = ""; // Clear previous choices
+    choicesContainer.innerHTML = "";
+
     if (dialogueQueue[dialogueIndex].choices) {
       for (const choice of dialogueQueue[dialogueIndex].choices) {
         const choiceButton = document.createElement("button");
         choiceButton.innerText = choice.text;
         choiceButton.addEventListener("click", (event) => {
-          event.stopPropagation(); // Stop event bubbling
+          event.stopPropagation();
           if (choice.action) {
             choice.action();
           }
@@ -36,8 +43,16 @@ const updateDialogue = () => {
     } else {
       attachContinueListeners();
     }
+
+    if (dialogueBox.parentElement) {
+      dialogueBox.parentElement.removeChild(dialogueBox);
+    }
+    parentElement.appendChild(dialogueBox);
   } else {
     dialogueBox.classList.remove("active");
+    if (dialogueBox.parentElement) {
+      dialogueBox.parentElement.removeChild(dialogueBox);
+    }
   }
 };
